Library Documentation

Documentation for NonlinearSchrodinger.jl's library.

Contents

Library Documentation

Contents

Index

Integrators

Functions

Index

NonlinearSchrodinger.Box

NonlinearSchrodinger.Calc

NonlinearSchrodinger.Operators

NonlinearSchrodinger.Sim

NonlinearSchrodinger.T1A!

NonlinearSchrodinger.T1A!

NonlinearSchrodinger.T1A_H!

NonlinearSchrodinger.T1B!

NonlinearSchrodinger.T1B!

NonlinearSchrodinger.T2A!

NonlinearSchrodinger.T2A!

NonlinearSchrodinger.T2A_H!

NonlinearSchrodinger.T2B!

NonlinearSchrodinger.T2B!

NonlinearSchrodinger.T4A_N!

NonlinearSchrodinger.T4A_N!

NonlinearSchrodinger.T4A_SF!

NonlinearSchrodinger.T4A_SF!

NonlinearSchrodinger.T4A_TJ!

NonlinearSchrodinger.T4A_TJ!

NonlinearSchrodinger.T4B_N!

NonlinearSchrodinger.T4B_N!

NonlinearSchrodinger.T4B_SF!

NonlinearSchrodinger.T4B_SF!

NonlinearSchrodinger.T4B_TJ!

NonlinearSchrodinger.T4B_TJ!

NonlinearSchrodinger.T6A_N!

NonlinearSchrodinger.T6A_N!

NonlinearSchrodinger.T6A_OP!

NonlinearSchrodinger.T6A_OP!

NonlinearSchrodinger.T6A_SF!

NonlinearSchrodinger.T6A_SF!

NonlinearSchrodinger.T6A_TJ!

NonlinearSchrodinger.T6A_TJ!

NonlinearSchrodinger.T6B_N!

NonlinearSchrodinger.T6B_N!

NonlinearSchrodinger.T6B_OP!

NonlinearSchrodinger.T6B_OP!

NonlinearSchrodinger.T6B_SF!

NonlinearSchrodinger.T6B_SF!

NonlinearSchrodinger.T6B_TJ!

NonlinearSchrodinger.T6B_TJ!

NonlinearSchrodinger.T8A_N!

NonlinearSchrodinger.T8A_N!

NonlinearSchrodinger.T8A_OP!

NonlinearSchrodinger.T8A_OP!

NonlinearSchrodinger.T8A_SF!

NonlinearSchrodinger.T8A_SF!

NonlinearSchrodinger.T8A_TJ!

NonlinearSchrodinger.T8A_TJ!

NonlinearSchrodinger.T8B_N!

NonlinearSchrodinger.T8B_N!

NonlinearSchrodinger.T8B_OP!

NonlinearSchrodinger.T8B_OP!

NonlinearSchrodinger.T8B_SF!

NonlinearSchrodinger.T8B_SF!

NonlinearSchrodinger.T8B_TJ!

NonlinearSchrodinger.T8B_TJ!

NonlinearSchrodinger.compute_IoM!

NonlinearSchrodinger.params

NonlinearSchrodinger.solve!

NonlinearSchrodinger.ψ₀_periodic

Integrators

NonlinearSchrodinger.T1A! — Function

T1A!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic first order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T1B! — Function

T1B!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic first order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T2A! — Function

T2A!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic second order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T2B! — Function

T2B!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic second order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4A_TJ! — Function

T4A_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Fourth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4B_TJ! — Function

T4B_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Fourth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6A_TJ! — Function

T6A_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Sixth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6B_TJ! — Function

T6B_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Sixth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8A_TJ! — Function

T8A_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Eighth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8B_TJ! — Function

T8A_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Eighth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4A_SF! — Function

T4A_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Fourth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4B_SF! — Function

T4B_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Fourth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6A_SF! — Function

T6A_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Sixth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6B_SF! — Function

T6B_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Sixth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8A_SF! — Function

T8A_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Eighth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8B_SF! — Function

T8B_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Eighth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4A_N! — Function

T4A_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Fourth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4B_N! — Function

T4B_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Fourth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6A_N! — Function

T6A_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Sixth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6B_N! — Function

T6B_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Sixth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8A_N! — Function

T8A_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Eighth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8B_N! — Function

T8B_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Eighth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6A_OP! — Function

T6A_OP!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using an Optimized Symplectic Sixth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6B_OP! — Function

T6B_OP!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using an Optimized Symplectic Sixth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8A_OP! — Function

T8A_OP!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using an Optimized Symplectic Eighth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8B_OP! — Function

T8B_OP!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using an Optimized Symplectic Eighth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

Functions

NonlinearSchrodinger.Box — Method

function Box(xᵣ::Pair, T; dx, Nₓ, Nₜ, n_periods)

Create a ::Box object with Nₜ transverse nodes, Nₓ longitudinal nodes (or dx grid spacing in the lontidunal direction, only one option can be specified), of size [(xᵣ.first), (xᵣ.second)] in the longitudinal direction and [(-T*n_periods/2), (T*n_periods/2))

NonlinearSchrodinger.Calc — Method

function Calc(λ::Array{Complex{TT}}, tₛ, xₛ, seed, box; m=0.0) where TT <: Real

Create a ::Calc object with eigenvalues λ, shifts xₛ and tₛ and seeding solution seed. box::Box is the Box object used for the calculation and m is the elliptic parameter for cnoidal solutions. seed can have the following values:

seed = "0" $\implies \psi_0 = 0$

seed = "exp" $\implies \psi_0 = e^{ix}$

seed = "dn" $\implies \psi_0 = dn(t, m)e^{ix(1-m/2)}$

seed = "cn" $\implies \psi_0 = \sqrt{m}cn(t, m)e^{ix(m - 1/2)}$

See also: Box

NonlinearSchrodinger.Operators — Method

function Operators(sim)

Construct appropriate operators for sim::Sim. These include the Fourier operator F̂, inverse Fourier operator F̃̂, dispersion operaotr K̂, and Hirota/Sasa-Satsuma operators B̂ and B̂B. This object also includes some arrays for temporary operations in Nystrom integrators.

See also: Sim

NonlinearSchrodinger.Sim — Method

function Sim(λ, box::Box, ψ₀::Array{Complex{TT}, 1}, T̂; α = 0.0, ϵ=0.0, β=0.0) where TT <: Rea

Create a ::Sim object corresponding to eigenvalue λ (only used for breathers, ignored for arbitrary initial conditions), initial condition ψ₀ and algorithm T̂. α is the Hirota equation parameter and ϵ is the Sasa-Satsuma equation parameter. β controls pruning.

See also: Box

NonlinearSchrodinger.T1A! — Method

T1A!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic first order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T1A_H! — Method

T₁ʰ(ψ, ω, dx, F)

Compute ψ', i.e. ψ advanced a step dx forward using a symplectic second order integrator. ψ' is defined on an FFT grid with frequencies ω using an FFT plan F. Do not call this explicitly and use solve! instead.

See also: solve!

NonlinearSchrodinger.T1B! — Method

T1B!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic first order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T2A! — Method

T2A!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic second order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T2A_H! — Method

T₂ʰ(ψ, ω, dx, F)

Compute ψ', i.e. ψ advanced a step dx forward using a symplectic second order integrator. ψ' is defined on an FFT grid with frequencies ω using an FFT plan F. Do not call this explicitly and use solve! instead.

See also: solve!

NonlinearSchrodinger.T2B! — Method

T2B!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic second order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4A_N! — Method

T4A_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Fourth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4A_SF! — Method

T4A_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Fourth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4A_TJ! — Method

T4A_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Fourth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4B_N! — Method

T4B_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Fourth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4B_SF! — Method

T4B_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Fourth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4B_TJ! — Method

T4B_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Fourth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6A_N! — Method

T6A_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Sixth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6A_OP! — Method

T6A_OP!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using an Optimized Symplectic Sixth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6A_SF! — Method

T6A_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Sixth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6A_TJ! — Method

T6A_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Sixth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6B_N! — Method

T6B_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Sixth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6B_OP! — Method

T6B_OP!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using an Optimized Symplectic Sixth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6B_SF! — Method

T6B_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Sixth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6B_TJ! — Method

T6B_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Sixth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8A_N! — Method

T8A_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Eighth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8A_OP! — Method

T8A_OP!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using an Optimized Symplectic Eighth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8A_SF! — Method

T8A_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Eighth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8A_TJ! — Method

T8A_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Eighth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8B_N! — Method

T8B_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Eighth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8B_OP! — Method

T8B_OP!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using an Optimized Symplectic Eighth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8B_SF! — Method

T8B_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Eighth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8B_TJ! — Method

T8A_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Eighth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.compute_IoM! — Method

function compute_IoM!(obj)

Computes the integrals of motion of obj.ψ and saves them in respective fields of obj. obj can be a ::Sim or ::Calc object. The fields are:

obj.KE  # Array containing the kinetic energy K(x)
obj.PE  # Array containing the potential energy V(x)
obj.E   # Array containing the energy H(x)
obj.N   # Array containing the norm N(x)
obj.P   # Array containing the momentum P(x)
obj.dE  # Array containing the energy error δE(x)
obj.dN  # Array containing the norm error δN(x)
obj.dP  # Array containing the momentum error δP(x)

Result can be plotted using plot(obj, :IoM)

NonlinearSchrodinger.params — Method

function params(; kwargs...)

Computes parameters

NonlinearSchrodinger.solve! — Method

solve!(sim::Sim)

Solves the Sim object sim.

See also: Sim

NonlinearSchrodinger.ψ₀_periodic — Method

function ψ₀_periodic(coeff::Array, box::Box, Ω; phase=0)

Computes an initial wavefunction for the box::Box, with fundamental frequency Ω and coefficients $A_{1\ldots n}$ = coeff and an overall phase $e^{i \phi t}$ where $\phi$ = phase. i.e. $\psi_0$ is of the form:

$\psi(x=0, t) = e^{i \phi t} (A_0 + 2 \sum_{1}^{n} A_m \cos(m \Omega t))$

where:

$A_0 = \sqrt{1 - 2 \sum_{m=1}^n |A_m|^2}$