Integrators

This page shows the availabe integrators. Integrators prefixed with _H are used for the Hirota equation, and those with _SS are for the Sasa-Satsuma equation. Note that both these equations are not fully tested and should be used with utmost care due to the periodic boundary conditions.

NonlinearSchrodinger.T1A!

NonlinearSchrodinger.T1A_H!

NonlinearSchrodinger.T1A_SS!

NonlinearSchrodinger.T1B!

NonlinearSchrodinger.T2A!

NonlinearSchrodinger.T2A_H!

NonlinearSchrodinger.T2B!

NonlinearSchrodinger.T4A_CMP!

NonlinearSchrodinger.T4A_SF!

NonlinearSchrodinger.T4A_TJ!

NonlinearSchrodinger.T4B_CMP!

NonlinearSchrodinger.T4B_SF!

NonlinearSchrodinger.T4B_TJ!

NonlinearSchrodinger.T6A_CMP!

NonlinearSchrodinger.T6A_KLs9!

NonlinearSchrodinger.T6A_SF!

NonlinearSchrodinger.T6A_Ss14!

NonlinearSchrodinger.T6A_TJ!

NonlinearSchrodinger.T6A_Ys7!

NonlinearSchrodinger.T6B_CMP!

NonlinearSchrodinger.T6B_KLs9!

NonlinearSchrodinger.T6B_SF!

NonlinearSchrodinger.T6B_Ss14!

NonlinearSchrodinger.T6B_TJ!

NonlinearSchrodinger.T6B_Ys7!

NonlinearSchrodinger.T8A_CMP!

NonlinearSchrodinger.T8A_KLs17!

NonlinearSchrodinger.T8A_SF!

NonlinearSchrodinger.T8A_Ss15!

NonlinearSchrodinger.T8A_TJ!

NonlinearSchrodinger.T8B_CMP!

NonlinearSchrodinger.T8B_KLs17!

NonlinearSchrodinger.T8B_SF!

NonlinearSchrodinger.T8B_Ss15!

NonlinearSchrodinger.T8B_TJ!

NonlinearSchrodinger.T1A! — Function

T1A!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic first order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T1B! — Function

T1B!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic first order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T2A! — Function

T2A!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic second order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T2B! — Function

T2B!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic second order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4A_TJ! — Function

T4A_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Fourth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4B_TJ! — Function

T4B_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Fourth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6A_TJ! — Function

T6A_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Sixth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6B_TJ! — Function

T6B_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Sixth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8A_TJ! — Function

T8A_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Eighth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8B_TJ! — Function

T8A_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Eighth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4A_SF! — Function

T4A_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Fourth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4B_SF! — Function

T4B_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Fourth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6A_SF! — Function

T6A_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Sixth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6B_SF! — Function

T6B_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Sixth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8A_SF! — Function

T8A_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Eighth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8B_SF! — Function

T8B_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Eighth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4A_CMP! — Function

T4A_CMP!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin Multi-Product Fourth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4B_CMP! — Function

T4B_CMP!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin Multi-Product Fourth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6A_CMP! — Function

T6A_CMP!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin Multi-Product Sixth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6B_CMP! — Function

T6B_CMP!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin Multi-Product Sixth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8A_CMP! — Function

T8A_CMP!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin Multi-Product Eighth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8B_CMP! — Function

T8B_CMP!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin Multi-Product Eighth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6A_Ss14! — Function

T6A_Ss14!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using Suzuki's s14 Symplectic Sixth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6B_Ss14! — Function

T6B_Ss14!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using Suzuki's s14 Symplectic Sixth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6A_Ys7! — Function

T6A_Ys7!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using Yoshida's s7 Symplectic Sixth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6B_Ys7! — Function

T6B_Ys7!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using Yoshida's s7 Symplectic Sixth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6A_KLs9! — Function

T6A_KLs9!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using Kahan & Li's s9 Symplectic Sixth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6B_KLs9! — Function

T6B_KLs9!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using Kahan & Li's s9 Symplectic Sixth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8A_Ss15! — Function

T8A_Ss15!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using Suzuki's s15 Symplectic Eighth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8B_Ss15! — Function

T8B_Ss15!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using Suzuki's s15 Symplectic Eighth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8A_KLs17! — Function

T8A_KLs17!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using Kahan & Li's s17 Symplectic Eighth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8B_KLs17! — Function

T8B_KLs17!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using Kahan & Li's s17 Symplectic Eighth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T1A_H! — Function

T1A_H!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic first order integrator of type A for the Hirota equation. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T2A_H! — Function

T2A_H!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic second order integrator of type A for the Hirota equation. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T1A_SS! — Function

T1A_SS!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic first order integrator of type A for the Sasa-Satsuma equation. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators