Documentation for NonlinearSchrodinger.jl's internals.

Contents

Contents

Index

API

Index

NonlinearSchrodinger.Operators

NonlinearSchrodinger.soln_loop_A

NonlinearSchrodinger.soln_loop_B

API

NonlinearSchrodinger.soln_loop_A — Method

soln_loop_A(sim, ops, ind_p))

Propagate in x for integrators of type A (i.e. starting with nonlinear step, thus in real space).

See also: Sim, solve!

NonlinearSchrodinger.soln_loop_B — Method

soln_loop_B(sim, ops, ind_p))

Propagate in x for integrators of type B (i.e. starting with dispersion step, thus in frequency space).

See also: Sim, solve!

NonlinearSchrodinger.Operators — Method

function Operators(sim)

Construct appropriate operators for sim::Sim. These include the Fourier operator F̂, inverse Fourier operator F̃̂, dispersion operaotr K̂, and Hirota/Sasa-Satsuma operators B̂ and B̂B. This object also includes some arrays for temporary operations in Nystrom integrators.

See also: Sim