Library Documentation

Documentation for NLSS.jl's library.

Contents

Library Documentation

Contents

Index

Simulation

Calculation

Plotter

Utilities

Index

NLSS.Plotter.plot_IoM

NLSS.Plotter.plot_ψ

NLSS.Plotter.plot_ψ̃

NLSS.Simulation.T₁ʰ

NLSS.Simulation.T₁ˢ

NLSS.Simulation.T₂ʰ

NLSS.Simulation.T₂ˢ

NLSS.Simulation.T₄ʰ

NLSS.Simulation.T₄ˢ

NLSS.Simulation.T₆ˢ

NLSS.Simulation.T₈ˢ

NLSS.Simulation.print

NLSS.Simulation.solve!

NLSS.Simulation.ψ₀_periodic

NLSS.Utilities.compute_IoM!

NLSS.Utilities.compute_spectrum!

Simulation

NLSS.Simulation.T₁ʰ — Method

T₁ʰ(ψ, ω, dx, F)

Compute ψ', i.e. ψ advanced a step dx forward using a symplectic second order integrator. ψ' is defined on an FFT grid with frequencies ω using an FFT plan F. Do not call this explicitly and use solve! instead.

See also: solve!

NLSS.Simulation.T₁ˢ — Method

T₁ˢ(ψ, ω, dx, F)

Compute ψ', i.e. ψ advanced a step dx forward using a symplectic second order integrator. ψ' is defined on an FFT grid with frequencies ω using an FFT plan F. Do not call this explicitly and use solve! instead.

See also: solve!

NLSS.Simulation.T₂ʰ — Method

T₂ʰ(ψ, ω, dx, F)

Compute ψ', i.e. ψ advanced a step dx forward using a symplectic second order integrator. ψ' is defined on an FFT grid with frequencies ω using an FFT plan F. Do not call this explicitly and use solve! instead.

See also: solve!

NLSS.Simulation.T₂ˢ — Method

T₂ˢ(ψ, ω, dx, F)

Compute ψ', i.e. ψ advanced a step dx forward using a symplectic second order integrator. ψ' is defined on an FFT grid with frequencies ω using an FFT plan F. Do not call this explicitly and use solve! instead.

See also: solve!

NLSS.Simulation.T₄ʰ — Method

T₄ʰ(ψ, ω, dx, F)

Compute ψ', i.e. ψ advanced a step dx forward using a symplectic fourth order integrator. ψ' is defined on an FFT grid with frequencies ω using an FFT plan F. Do not call this explicitly and use solve! instead.

See also: solve!, T2

NLSS.Simulation.T₄ˢ — Method

T₄ˢ(ψ, ω, dx, F)

Compute ψ', i.e. ψ advanced a step dx forward using a symplectic fourth order integrator. ψ' is defined on an FFT grid with frequencies ω using an FFT plan F. Do not call this explicitly and use solve! instead.

See also: solve!, T2

NLSS.Simulation.T₆ˢ — Method

T₆ˢ(ψ, ω, dx, F)

Compute ψ', i.e. ψ advanced a step dx forward using a symplectic sixth order integrator. ψ' is defined on an FFT grid with frequencies ω using an FFT plan F. Do not call this explicitly and use solve! instead.

See also: solve!, T₄ˢ

NLSS.Simulation.T₈ˢ — Method

T₈ˢ(ψ, ω, dx, F)

Compute ψ', i.e. ψ advanced a step dx forward using a symplectic eighth order integrator. ψ' is defined on an FFT grid with frequencies ω using an FFT plan F. Do not call this explicitly and use solve! instead.

See also: solve!, T₆ˢ

NLSS.Simulation.print — Method

function print(sim::Sim)

Prints information about the Sim instance sim

NLSS.Simulation.solve! — Method

solve!(sim::Simulation)

Solves the Simulation object sim using the techniques its attributes specify.

See also: init_sim, NLSS.Plotter.plot_ψ

NLSS.Simulation.ψ₀_periodic — Method

function ψ₀_periodic(coeff, box::SimBox, params::SimParamseters; phase=0)

Computes an initial wavefunction for the SimBox box, with fundamental frequency sim.Ω and coefficients $A_1...n$ = coeff and an overall phase exp(i phase t), i.e. of the form:

TODO: insert latex form here

See also: init_sim

Calculation

Plotter

NLSS.Plotter.plot_IoM — Function

function plot_IoM(sim::Sim, x_res::Int64 = 500)

Plots the integrals of motion for a Sim object sim with a resolution x_res points in the x-direction. Produces plots of the energy, kinetic energy, potential energy, energy error, particle number and momentum

See also: Simulation.compute_CoM!

NLSS.Plotter.plot_ψ — Method

plot_ψ(sim::Sim; mode = "density", power=1, x_res=500, t_res=512)

Plot sim.abs.(ψ).^power on a grid of size approximately t_resxx_res The function can plot either a heatmap (mode = "density") or a 3D surface (mode = "surface")

TODO: Insert pictures here

See also: Simulation.solve!

NLSS.Plotter.plot_ψ̃ — Method

plot_ψ̃(sim::Sim; mode = "density", power=1, x_res=500, t_res=512)

Plot log(sim.abs.(ψ)) on a grid of size approximately ω_resxx_res if mode = density, or with n_lines total, skipping skip between each line if mode = lines. The latter uses x_res points as well.

TODO: Insert pictures here

See also: Simulation.compute_spectrum!

Utilities

NLSS.Utilities.compute_IoM! — Method

function compute_IoM!(obj)

Computes the integrals of motion of obj.ψ and saves them in respective fields of obj. obj can be a Sim or Calc object

See also: NLSS.Plotter.plot_CoM

NLSS.Utilities.compute_spectrum! — Method

function compute_spectrum!(obj)

Computes the normalized spectrum of obj.ψ with the center frequency shifted to the center and saves it in obj.ψ̃. obj can be a Sim or Calc object

See also: NLSS.Plotter.plot_ψ̃