Library Documentation

Documentation for NLSS.jl's library.

Contents

Library Documentation

Contents

Index

Test

Index

NonlinearSchrodinger.plot_IoM

NonlinearSchrodinger.plot_ψ

NonlinearSchrodinger.plot_ψ̃

NonlinearSchrodinger.solve!

Test

NonlinearSchrodinger.plot_IoM — Method

function plot_IoM(sim::Sim, x_res::Int64 = 500)

Plots the integrals of motion for a Sim object sim with a resolution x_res points in the x-direction. Produces plots of the energy, kinetic energy, potential energy, energy error, particle number and momentum

See also: Simulation.compute_CoM!

NonlinearSchrodinger.plot_ψ — Method

plot_ψ(sim::Sim; mode = "density", power=1, x_res=500, t_res=512)

Plot sim.abs.(ψ).^power on a grid of size approximately t_resxx_res The function can plot either a heatmap (mode = "density") or a 3D surface (mode = "surface")

TODO: Insert pictures here

See also: Simulation.solve!

NonlinearSchrodinger.plot_ψ̃ — Method

plot_ψ̃(sim::Sim; mode = "density", power=1, x_res=500, t_res=512)

Plot log(sim.abs.(ψ)) on a grid of size approximately ω_resxx_res if mode = density, or with n_lines total, skipping skip between each line if mode = lines. The latter uses x_res points as well.

TODO: Insert pictures here

See also: Simulation.compute_spectrum!

NonlinearSchrodinger.solve! — Method

solve!(sim::Simulation)

Solves the Simulation object sim using the techniques its attributes specify.

See also: init_sim, NLSS.Plotter.plot_ψ