Library Documentation

Documentation for NonlinearSchrodinger.jl's library.

Contents

Library Documentation

Contents

Index

Functions

Index

NonlinearSchrodinger.Box

NonlinearSchrodinger.Calc

NonlinearSchrodinger.Operators

NonlinearSchrodinger.Sim

NonlinearSchrodinger.compute_IoM!

NonlinearSchrodinger.params

NonlinearSchrodinger.solve!

NonlinearSchrodinger.ψ₀_periodic

Functions

NonlinearSchrodinger.solve! — Method

solve!(sim::Sim)

Solves the Sim object sim.

See also: Sim

NonlinearSchrodinger.Box — Method

function Box(xᵣ::Pair, T; dx, Nₓ, Nₜ, n_periods)

Create a ::Box object with Nₜ transverse nodes, Nₓ longitudinal nodes (or dx grid spacing in the lontidunal direction, only one option can be specified), of size [(xᵣ.first), (xᵣ.second)] in the longitudinal direction and [(-T*n_periods/2), (T*n_periods/2))

NonlinearSchrodinger.Calc — Method

function Calc(λ::Array{Complex{TT}}, tₛ, xₛ, seed, box; m=0.0) where TT <: Real

Create a ::Calc object with eigenvalues λ, shifts xₛ and tₛ and seeding solution seed. box::Box is the Box object used for the calculation and m is the elliptic parameter for cnoidal solutions. seed can have the following values:

seed = "0" $\implies \psi_0 = 0$

seed = "exp" $\implies \psi_0 = e^{ix}$

seed = "dn" $\implies \psi_0 = dn(t, m)e^{ix(1-m/2)}$

seed = "cn" $\implies \psi_0 = \sqrt{m}cn(t, m)e^{ix(m - 1/2)}$

See also: Box

NonlinearSchrodinger.Operators — Method

function Operators(sim)

Construct appropriate operators for sim::Sim. These include the Fourier operator F̂, inverse Fourier operator F̃̂, dispersion operaotr K̂, and Hirota/Sasa-Satsuma operators B̂ and B̂B. This object also includes some arrays for temporary operations in Nystrom integrators.

See also: Sim

NonlinearSchrodinger.Sim — Method

function Sim(λ, box::Box, ψ₀::Array{Complex{TT}, 1}, T̂; α = 0.0, ϵ=0.0, β=0.0) where TT <: Rea

Create a ::Sim object corresponding to eigenvalue λ (only used for breathers, ignored for arbitrary initial conditions), initial condition ψ₀ and algorithm T̂. α is the Hirota equation parameter and ϵ is the Sasa-Satsuma equation parameter. β controls pruning.

See also: Box

NonlinearSchrodinger.compute_IoM! — Method

function compute_IoM!(obj)

Computes the integrals of motion of obj.ψ and saves them in respective fields of obj. obj can be a ::Sim or ::Calc object. The fields are:

obj.KE  # Array containing the kinetic energy K(x)
obj.PE  # Array containing the potential energy V(x)
obj.E   # Array containing the energy H(x)
obj.N   # Array containing the norm N(x)
obj.P   # Array containing the momentum P(x)
obj.dE  # Array containing the energy error δE(x)
obj.dN  # Array containing the norm error δN(x)
obj.dP  # Array containing the momentum error δP(x)

Result can be plotted using plot(obj, :IoM)

NonlinearSchrodinger.params — Method

function params(; kwargs...)

Computes parameters

NonlinearSchrodinger.ψ₀_periodic — Method

function ψ₀_periodic(coeff::Array, box::Box, Ω; phase=0)

Computes an initial wavefunction for the box::Box, with fundamental frequency Ω and coefficients $A_{1\ldots n}$ = coeff and an overall phase $e^{i \phi t}$ where $\phi$ = phase. i.e. $\psi_0$ is of the form:

$\psi(x=0, t) = e^{i \phi t} (A_0 + 2 \sum_{1}^{n} A_m \cos(m \Omega t))$

where:

$A_0 = \sqrt{1 - 2 \sum_{m=1}^n |A_m|^2}$