Library Documentation

Documentation for NonlinearSchrodinger.jl's library.

Contents

Library Documentation

Contents

Index

Integrators

Functions

Index

NonlinearSchrodinger.T1A!

NonlinearSchrodinger.T1A!

NonlinearSchrodinger.T1A_H!

NonlinearSchrodinger.T1B!

NonlinearSchrodinger.T1B!

NonlinearSchrodinger.T2A!

NonlinearSchrodinger.T2A!

NonlinearSchrodinger.T2A_H!

NonlinearSchrodinger.T2B!

NonlinearSchrodinger.T2B!

NonlinearSchrodinger.T4A_N!

NonlinearSchrodinger.T4A_N!

NonlinearSchrodinger.T4A_SF!

NonlinearSchrodinger.T4A_SF!

NonlinearSchrodinger.T4A_TJ!

NonlinearSchrodinger.T4A_TJ!

NonlinearSchrodinger.T4B_N!

NonlinearSchrodinger.T4B_N!

NonlinearSchrodinger.T4B_SF!

NonlinearSchrodinger.T4B_SF!

NonlinearSchrodinger.T4B_TJ!

NonlinearSchrodinger.T4B_TJ!

NonlinearSchrodinger.T6A_N!

NonlinearSchrodinger.T6A_N!

NonlinearSchrodinger.T6A_OP!

NonlinearSchrodinger.T6A_OP!

NonlinearSchrodinger.T6A_SF!

NonlinearSchrodinger.T6A_SF!

NonlinearSchrodinger.T6A_TJ!

NonlinearSchrodinger.T6A_TJ!

NonlinearSchrodinger.T6B_N!

NonlinearSchrodinger.T6B_N!

NonlinearSchrodinger.T6B_OP!

NonlinearSchrodinger.T6B_OP!

NonlinearSchrodinger.T6B_SF!

NonlinearSchrodinger.T6B_SF!

NonlinearSchrodinger.T6B_TJ!

NonlinearSchrodinger.T6B_TJ!

NonlinearSchrodinger.T8A_N!

NonlinearSchrodinger.T8A_N!

NonlinearSchrodinger.T8A_OP!

NonlinearSchrodinger.T8A_OP!

NonlinearSchrodinger.T8A_SF!

NonlinearSchrodinger.T8A_SF!

NonlinearSchrodinger.T8A_TJ!

NonlinearSchrodinger.T8A_TJ!

NonlinearSchrodinger.T8B_N!

NonlinearSchrodinger.T8B_N!

NonlinearSchrodinger.T8B_OP!

NonlinearSchrodinger.T8B_OP!

NonlinearSchrodinger.T8B_SF!

NonlinearSchrodinger.T8B_SF!

NonlinearSchrodinger.T8B_TJ!

NonlinearSchrodinger.T8B_TJ!

NonlinearSchrodinger.compute_IoM!

NonlinearSchrodinger.compute_spectrum!

NonlinearSchrodinger.params

NonlinearSchrodinger.print

NonlinearSchrodinger.solve!

NonlinearSchrodinger.ψ₀_periodic

Integrators

NonlinearSchrodinger.T1A! — Function

T1A!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic first order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T1B! — Function

T1B!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic first order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T2A! — Function

T2A!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic second order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T2B! — Function

T2B!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic second order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4A_TJ! — Function

T4A_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Fourth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4B_TJ! — Function

T4B_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Fourth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6A_TJ! — Function

T6A_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Sixth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6B_TJ! — Function

T6B_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Sixth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8A_TJ! — Function

T8A_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Eighth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8B_TJ! — Function

T8A_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Eighth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4A_SF! — Function

T4A_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Fourth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4B_SF! — Function

T4B_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Fourth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6A_SF! — Function

T6A_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Sixth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6B_SF! — Function

T6B_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Sixth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8A_SF! — Function

T8A_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Eighth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8B_SF! — Function

T8B_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Eighth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4A_N! — Function

T4A_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Fourth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4B_N! — Function

T4B_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Fourth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6A_N! — Function

T6A_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Sixth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6B_N! — Function

T6B_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Sixth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8A_N! — Function

T8A_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Eighth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8B_N! — Function

T8B_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Eighth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6A_OP! — Function

T6A_OP!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using an Optimized Symplectic Sixth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6B_OP! — Function

T6B_OP!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using an Optimized Symplectic Sixth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8A_OP! — Function

T8A_OP!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using an Optimized Symplectic Eighth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8B_OP! — Function

T8B_OP!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using an Optimized Symplectic Eighth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

Functions

NonlinearSchrodinger.T1A! — Method

T1A!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic first order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T1A_H! — Method

T₁ʰ(ψ, ω, dx, F)

Compute ψ', i.e. ψ advanced a step dx forward using a symplectic second order integrator. ψ' is defined on an FFT grid with frequencies ω using an FFT plan F. Do not call this explicitly and use solve! instead.

See also: solve!

NonlinearSchrodinger.T1B! — Method

T1B!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic first order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T2A! — Method

T2A!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic second order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T2A_H! — Method

T₂ʰ(ψ, ω, dx, F)

Compute ψ', i.e. ψ advanced a step dx forward using a symplectic second order integrator. ψ' is defined on an FFT grid with frequencies ω using an FFT plan F. Do not call this explicitly and use solve! instead.

See also: solve!

NonlinearSchrodinger.T2B! — Method

T2B!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic second order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4A_N! — Method

T4A_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Fourth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4A_SF! — Method

T4A_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Fourth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4A_TJ! — Method

T4A_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Fourth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4B_N! — Method

T4B_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Fourth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4B_SF! — Method

T4B_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Fourth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4B_TJ! — Method

T4B_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Fourth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6A_N! — Method

T6A_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Sixth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6A_OP! — Method

T6A_OP!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using an Optimized Symplectic Sixth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6A_SF! — Method

T6A_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Sixth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6A_TJ! — Method

T6A_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Sixth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6B_N! — Method

T6B_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Sixth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6B_OP! — Method

T6B_OP!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using an Optimized Symplectic Sixth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6B_SF! — Method

T6B_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Sixth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6B_TJ! — Method

T6B_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Sixth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8A_N! — Method

T8A_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Eighth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8A_OP! — Method

T8A_OP!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using an Optimized Symplectic Eighth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8A_SF! — Method

T8A_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Eighth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8A_TJ! — Method

T8A_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Eighth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8B_N! — Method

T8B_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Eighth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8B_OP! — Method

T8B_OP!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using an Optimized Symplectic Eighth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8B_SF! — Method

T8B_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Eighth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8B_TJ! — Method

T8A_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Eighth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.compute_IoM! — Method

function compute_IoM!(obj)

Computes the integrals of motion of obj.ψ and saves them in respective fields of obj. obj can be a Sim or Calc object

See also: NLSS.Plotter.plot_CoM

NonlinearSchrodinger.compute_spectrum! — Method

function compute_spectrum!(obj)

Computes the normalized spectrum of obj.ψ with the center frequency shifted to the center and saves it in obj.ψ̃. obj can be a Sim or Calc object

See also: NLSS.Plotter.plot_ψ̃

NonlinearSchrodinger.params — Method

function params(; kwargs...)

Computes parameters

NonlinearSchrodinger.print — Method

function print(sim::Sim)

Prints information about the Sim instance sim

NonlinearSchrodinger.solve! — Method

solve!(sim::Simulation)

Solves the Simulation object sim using the techniques its attributes specify.

See also: init_sim, NLSS.Plotter.plot_ψ

NonlinearSchrodinger.ψ₀_periodic — Method

function ψ₀_periodic(coeff, box::SimBox, params::SimParamseters; phase=0)

Computes an initial wavefunction for the SimBox box, with fundamental frequency sim.Ω and coefficients $A_1...n$ = coeff and an overall phase exp(i phase t), i.e. of the form:

TODO: insert latex form here

See also: init_sim