Integrators

This page shows the availabe integrators.

NonlinearSchrodinger.T1A!

NonlinearSchrodinger.T1A_H!

NonlinearSchrodinger.T1A_SS!

NonlinearSchrodinger.T1B!

NonlinearSchrodinger.T2A!

NonlinearSchrodinger.T2A_H!

NonlinearSchrodinger.T2B!

NonlinearSchrodinger.T4A_N!

NonlinearSchrodinger.T4A_SF!

NonlinearSchrodinger.T4A_TJ!

NonlinearSchrodinger.T4B_N!

NonlinearSchrodinger.T4B_SF!

NonlinearSchrodinger.T4B_TJ!

NonlinearSchrodinger.T6A_N!

NonlinearSchrodinger.T6A_OP!

NonlinearSchrodinger.T6A_SF!

NonlinearSchrodinger.T6A_TJ!

NonlinearSchrodinger.T6B_N!

NonlinearSchrodinger.T6B_OP!

NonlinearSchrodinger.T6B_SF!

NonlinearSchrodinger.T6B_TJ!

NonlinearSchrodinger.T8A_N!

NonlinearSchrodinger.T8A_OP!

NonlinearSchrodinger.T8A_SF!

NonlinearSchrodinger.T8A_TJ!

NonlinearSchrodinger.T8B_N!

NonlinearSchrodinger.T8B_OP!

NonlinearSchrodinger.T8B_SF!

NonlinearSchrodinger.T8B_TJ!

NonlinearSchrodinger.T1A! — Function

T1A!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic first order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T1B! — Function

T1B!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic first order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T2A! — Function

T2A!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic second order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T2B! — Function

T2B!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic second order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4A_TJ! — Function

T4A_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Fourth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4B_TJ! — Function

T4B_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Fourth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6A_TJ! — Function

T6A_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Sixth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6B_TJ! — Function

T6B_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Sixth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8A_TJ! — Function

T8A_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Eighth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8B_TJ! — Function

T8A_TJ!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Triple Jump Eighth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4A_SF! — Function

T4A_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Fourth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4B_SF! — Function

T4B_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Fourth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6A_SF! — Function

T6A_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Sixth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6B_SF! — Function

T6B_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Sixth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8A_SF! — Function

T8A_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Eighth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8B_SF! — Function

T8B_SF!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic Suzuki's Fractal Eighth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4A_N! — Function

T4A_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Fourth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T4B_N! — Function

T4B_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Fourth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6A_N! — Function

T6A_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Sixth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6B_N! — Function

T6B_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Sixth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8A_N! — Function

T8A_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Eighth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8B_N! — Function

T8B_N!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a Chin-Nystrom Eighth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6A_OP! — Function

T6A_OP!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using an Optimized Symplectic Sixth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T6B_OP! — Function

T6B_OP!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using an Optimized Symplectic Sixth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8A_OP! — Function

T8A_OP!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using an Optimized Symplectic Eighth order integrator of type A. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T8B_OP! — Function

T8B_OP!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using an Optimized Symplectic Eighth order integrator of type B. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T1A_H! — Function

T1A_H!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic first order integrator of type A for the Hirota equation. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T2A_H! — Function

T2A_H!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic second order integrator of type A for the Hirota equation. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators

NonlinearSchrodinger.T1A_SS! — Function

T1A_SS!(ψₒ, ψᵢ, dx, ops)

Compute ψₒ, i.e. ψᵢ advanced a step dx forward using a symplectic first order integrator of type A for the Sasa-Satsuma equation. The structure ops::Operators contains the FFT plans and the kinetic energy operators.

See also: solve!, Operators